pkg_managers: gomod: Add preliminary support for Go 1.21
One of Go 1.21's main features is specifying the suggested version of
toolchain to be used for compiling modules as a means of forwards
compatibility.
However, this patch only adds preliminary support for Go 1.21 in that
it essentially explicitly denies usage of toolchains (which is fine,
because those versions are only suggestive, Go always prefers building
with the bundled toolchain anyway) by the means of explicitly setting
the GOTOOLCHAIN env variable to 'local' which instructs Go to always
use it bundled toolchain (provided it's new enough to conform to the
minimum required version of Go indicated by the 'go' line). This patch
makes sure we set GOTOOLCHAIN=local during fetching of the dependencies
as well as for user container builds.

STONEBLD-2047

Signed-off-by: Erik Skultety <eskultet@redhat.com>